pubsub_end_to_end/requirements.txt
google-cloud-speech==2.0.0
spacy==2.2.0
requests==2.25.1
redis
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
